chatbot-frontend/backend/pyproject.toml
[project]
name = "marketing-strategy-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "groq>=0.37.1",
    "langchain>=1.2.6",
    "langchain-classic>=1.0.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.6",
    "passlib[bcrypt]>=1.7.4",
    "pip>=25.3",
    "pydantic>=2.12.5",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "supabase>=2.27.2",
    "tavily-python>=0.7.19",
    "uvicorn>=0.40.0",
]
